requirements.txt
tensorflow==2.19.0
tensorflow-hub==0.16.1
numpy==1.26.4
matplotlib==3.8.4
pillow==10.2.0
streamlit==1.29.0
